@@ -6,72 +6,48 @@
   "repository": "hfour/wsrun",
   "author": "hfour",
   "license": "MIT",
-  "jest": {
-    "verbose": false,
-    "preset": "ts-jest",
-    "testPathIgnorePatterns": [
-      "<rootDir>/tmp",
-      "<rootDir>/node_modules/",
-      "\\.util\\.ts$"
-    ],
-    "testMatch": [
-      "<rootDir>/tests/**/*.ts",
-      "<rootDir>/src/**/*.spec.ts"
-    ],
-    "modulePathIgnorePatterns": [
-      "<rootDir>/node_modules/",
-      "<rootDir>/tmp"
-    ]
-  },
   "bin": {
-    "wsrun": "./bin/wsrun.js"
+    "wsrun": "./bin/wsrun.mjs"
   },
   "files": [
     "bin/*",
     "build/**/!(*.spec.js|*.spec.js.map)"
   ],
   "devDependencies": {
-    "@types/bluebird": "^3.5.18",
-    "@types/glob": "^5.0.33",
-    "@types/jest": "^21.1.6",
-    "@types/lodash": "^4.14.85",
-    "@types/minimatch": "^3.0.3",
-    "@types/mkdirp": "^0.5.2",
-    "@types/mz": "^0.0.32",
-    "@types/node": "^8.0.53",
-    "@types/rimraf": "^2.0.2",
+    "tsup": "^8.0.2",
+    "@types/glob": "^8.1.0",
+    "@types/minimatch": "^5.1.2",
+    "@types/mkdirp": "^2.0.0",
+    "@types/node": "^20.12.8",
+    "@types/rimraf": "^4.0.5",
     "@types/split": "^0.3.28",
     "@types/yargs": "^13.0.0",
-    "jest": "^23.0.0",
-    "mkdirp": "^0.5.1",
-    "mz": "^2.7.0",
-    "npm-run-all": "^4.1.3",
-    "rimraf": "^2.6.2",
-    "semantic-release": "^15.13.18",
-    "ts-jest": "23.10.5",
-    "ts-mockito": "^2.5.0",
-    "typescript": "^3.1.1",
-    "prettier": "^1.19.1"
+    "mkdirp": "^3.0.1",
+    "npm-run-all": "^4.1.5",
+    "rimraf": "^5.0.5",
+    "semantic-release": "^23.0.8",
+    "prettier": "^3.2.5",
+    "ts-mockito": "^2.6.1",
+    "typescript": "^5.4.5",
+    "bun": "^1.1.7"
   },
   "scripts": {
-    "build": "tsc",
-    "watch": "tsc -w",
-    "test": "yarn build && yarn test:prettier && jest",
+    "build": "tsup src/index.ts --format esm --out-dir build",
+    "watch": "tsup --watch src/index.ts --format esm --out-dir build",
+    "test": "bun run build && bun test:prettier && bun test",
     "test:prettier": "prettier -c '**/*.ts' '**/*.json'",
-    "test:watch": "jest --watch",
+    "test:watch": "bun test --watch",
     "dev": "run-p test:watch watch",
-    "prepublish": "tsc",
+    "prepublish": "bun run build",
     "release": "semantic-release"
   },
   "dependencies": {
-    "bluebird": "^3.5.1",
-    "chalk": "^2.3.0",
-    "glob": "^7.1.2",
-    "lodash": "^4.17.4",
-    "minimatch": "^3.0.4",
+    "chalk": "^5.3.0",
+    "glob": "^10.3.12",
+    "minimatch": "^9.0.4",
     "split": "^1.0.1",
-    "throat": "^4.1.0",
-    "yargs": "^13.0.0",
-    "jest-changed-files": "^24.9.0"
+    "throat": "^6.0.2",
+    "yargs": "^17.7.2",
+    "jest-changed-files": "^29.7.0"
   }
 }
